Remove logic layer and make second cluster impl
