+++ src/app/clusters/general-diagnostics-server/BUILD.gn
@@ -17,8 +17,7 @@ import("//build_overrides/chip.gni")
 source_set("general-diagnostics-server") {
   sources = [
     "general-diagnostics-cluster.h",
-    "general-diagnostics-logic.cpp",
-    "general-diagnostics-logic.h",
+    "general-diagnostics-cluster.cpp",
   ]
 
   public_deps = [
