+++ src/app/clusters/general-diagnostics-server/app_config_dependent_sources.cmake
@@ -25,6 +25,5 @@ TARGET_SOURCES(
   ${APP_TARGET}
   PRIVATE
     "${CLUSTER_DIR}/general-diagnostics-cluster.h"
-    "${CLUSTER_DIR}/general-diagnostics-logic.h"
-    "${CLUSTER_DIR}/general-diagnostics-logic.cpp"
+    "${CLUSTER_DIR}/general-diagnostics-cluster.cpp"
 )
